Refactored name of n4 method file
